@@ -1,16 +1,18 @@
 {
   "name": "the-human-tech-blog-react",
   "private": true,
-  "version": "0.0.0",
+  "version": "1.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc -b && vite build",
+    "lint": "eslint .",
     "preview": "vite preview",
-    "lint": "eslint \"**/*.{ts,tsx,js,jsx}\" --max-warnings=0",
-    "format": "prettier --write .",
-    "commit": "cz",
-    "postinstall": "husky install"
+    "prepare": "if [ -f pnpm-lock.yaml ]; then echo '🚫 PNPM is not allowed in this project. Use NPM only.' && exit 1; fi"
+  },
+  "engines": {
+     "node": ">=20.18",
+     "npm": ">=10"
   },
   "dependencies": {
     "@tiptap/extension-image": "^2.12.0",
@@ -24,8 +26,8 @@
     "lucide-react": "^0.511.0",
     "marked": "^15.0.11",
     "path": "^0.12.7",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-google-recaptcha": "^3.1.0",
     "react-hook-form": "^7.56.4",
     "react-hot-toast": "^2.5.2",
@@ -41,35 +43,20 @@
     "@eslint/js": "^9.22.0",
     "@types/fs-extra": "^11.0.4",
     "@types/node": "^22.14.1",
-    "@types/react": "^19.0.10",
-    "@types/react-dom": "^19.0.4",
+    "@types/react": "^19.1.4",
+    "@types/react-dom": "^19.1.5",
     "@types/react-google-recaptcha": "^2.1.9",
     "@types/socket.io-client": "^3.0.0",
     "@vitejs/plugin-react": "^4.3.4",
     "commitizen": "^4.3.1",
-    "cz-conventional-changelog": "^3.3.0",
     "eslint": "^9.22.0",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^16.0.0",
-    "husky": "^9.1.7",
     "lint-staged": "^16.0.0",
-    "prettier": "^3.2.5",
     "sass": "^1.87.0",
     "typescript": "~5.7.2",
     "typescript-eslint": "^8.26.1",
     "vite": "^6.3.1"
-  },
-  "lint-staged": {
-    "**/*.{ts,tsx,js,jsx,json,scss,md}": [
-      "eslint --fix",
-      "prettier --write"
-    ]
-  },
-  "config": {
-    "commitizen": {
-      "path": "cz-conventional-changelog"
-    }
-  },
-  "packageManager": "pnpm@10.11.0+sha512.6540583f41cc5f628eb3d9773ecee802f4f9ef9923cc45b69890fb47991d4b092964694ec3a4f738a420c918a333062c8b925d312f42e4f0c263eb603551f977"
+  }
 }
